Cast inputs from uint8 to float before computing metrics

+++ token_bench/metrics_cli.py
@@ -164,8 +164,8 @@ def main_psnr_ssim() -> None:
         assert (
             input0_file.split("/")[-1] == input1_file.split("/")[-1]
         ), "file names must match"
-        input0 = read_video(input0_file)
-        input1 = read_video(input1_file)
+        input0 = read_video(input0_file).astype(np.float32)
+        input1 = read_video(input1_file).astype(np.float32)
 
         name = input0_file.split("/")[-1]
         psnr_value = PSNR(input0, input1)
